Refactored and polished Sensor Service related files.
